Move helpers
Move some helpers into the helpers file. Do some formatting.
